Set version to 0.1.3 for release build
Originally used version 1.0.0 - 1.0.2 as this was a mod
which I had no intentions of releasing but have decided
to move the major version to 0 to create an beta release
version for others to download and use with no expectation
for stability.

Because version 1.0.0 - 1.0.2 were never officially
released publicly I have no concerns about moving the
version backwards at risk of conflicting in future.