Fix returnSpectators for 1.19.3
The previous game mode doesn't seem to change in 1.19.3 as
it did in the previous version, where it now sets the value
of the previous game mode to itself for some reason.

Injection point is moved to before the game mode field is
updated and uses the field and method argument to instead
determine what game mode is being changed from and to.